Make remote call errors configurable
Resolves https://github.com/apple/swift-distributed-actors/issues/932
