+++ Sources/DistributedActors/ClusterSystem.swift
@@ -1207,7 +1207,7 @@ extension ClusterSystem {
     {
         let callID = UUID()
 
-        let timeout = RemoteCall.timeout ?? self.settings.defaultRemoteCallTimeout
+        let timeout = RemoteCall.timeout ?? self.settings.remoteCall.defaultTimeout
         let timeoutTask: Task<Void, Error> = Task.detached {
             try await Task.sleep(nanoseconds: UInt64(timeout.nanoseconds))
             guard !Task.isCancelled else {
@@ -1493,11 +1493,25 @@ public struct ClusterInvocationResultHandler: DistributedTargetInvocationResultH
             system.log.debug("Result handler, onThrow: \(error)", metadata: ["call/id": "\(callID)"])
             let reply: RemoteCallReply<_Done>
             if let codableError = error as? (Error & Codable) {
-                reply = .init(callID: callID, error: codableError)
+                switch system.settings.remoteCall.codableErrorAllowance {
+                case .custom(let allowedTypes) where errorTypeAllowed(error: error, allowedTypes: allowedTypes):
+                    reply = .init(callID: callID, error: codableError)
+                case .all: // compiler gets confused if this is grouped together with above
+                    reply = .init(callID: callID, error: codableError)
+                default:
+                    reply = .init(callID: callID, error: GenericRemoteCallError(message: "Remote call error of [\(type(of: error as Any))] type occurred"))
+                }
             } else {
                 reply = .init(callID: callID, error: GenericRemoteCallError(message: "Remote call error of [\(type(of: error as Any))] type occurred"))
             }
             try await channel.writeAndFlush(TransportEnvelope(envelope: Payload(payload: .message(reply)), recipient: recipient))
+        }
+
+        func errorTypeAllowed<Err>(error: Err, allowedTypes: [(Error & Codable).Type]) -> Bool {
+            func isErrorOfType<Err, T: Error & Codable>(_ error: Err, _: T.Type) -> Bool {
+                error is T // `is` and `as` don't work with stored type
+            }
+            return allowedTypes.contains { type in isErrorOfType(error, type) }
         }
     }
 }
