+++ Sources/DistributedActors/ClusterSystemSettings.swift
@@ -167,11 +167,8 @@ public struct ClusterSystemSettings {
     // ==== ------------------------------------------------------------------------------------------------------------
     // MARK: Distributed Actor Calls
 
-    /// If no other timeout is specified, this timeout is applied to every distributed call.
-    /// A "distributed call" is any function call of a distributed function on a 'remote' distributed actor.
-    ///
-    /// Set to `.effectivelyInfinite` to avoid setting a timeout, although this is not recommended.
-    public var defaultRemoteCallTimeout: Duration = .seconds(5)
+    /// A "remote call" is any function call of a `distributed` function on a 'remote' distributed actor.
+    public var remoteCall: RemoteCallSettings = .default
 
     // ==== ------------------------------------------------------------------------------------------------------------
     // MARK: TLS & Security settings
@@ -447,5 +444,34 @@ public struct ServiceDiscoverySettings {
     enum ServiceDiscoveryImplementation {
         case `static`(Set<Node>)
         case dynamic(AnyServiceDiscovery)
+    }
+}
+
+// ==== ----------------------------------------------------------------------------------------------------------------
+// MARK: Remote Call Settings
+
+extension ClusterSystemSettings {
+    public struct RemoteCallSettings {
+        public static var `default`: RemoteCallSettings {
+            .init()
+        }
+
+        /// If no other timeout is specified, this timeout is applied to every distributed call.
+        ///
+        /// Set to `.effectivelyInfinite` to avoid setting a timeout, although this is not recommended.
+        public var defaultTimeout: Duration = .seconds(5)
+
+        public var codableErrorAllowance: CodableErrorAllowance = .all
+
+        public enum CodableErrorAllowance {
+            /// All ``Codable`` errors will be converted to ``GenericRemoteCallError``.
+            case none
+
+            /// All ``Codable`` errors will be returned as-is.
+            case all
+
+            /// Only the indicated ``Codable`` errors are allowed. Others are converted to ``GenericRemoteCallError``.
+            case custom([(Error & Codable).Type])
+        }
     }
 }
